+++ mips/lib/mplinsolve.m
@@ -15,27 +15,38 @@ function [x, info] = mplinsolve(A, b, solver, opt)
 %           ''  - use default solver, currently this is
 %                   always the built-in backslash operator
 %           '\' - built-in backslash operator
-%           'LU' - use explicit default LU decomposition and back substitution
-%           'LU_AMD' - use LU decomposition with approximate minimum degree
-%                     (AMD) reordering
-%           'LU_GP' - use Gilbert-Peierls algorithm for LU Decomposition with
-%                  (AMD) reordering (e.g. ideal for power flow Jacobian)
+%           'LU' - use explicit LU decomposition and back substitution
+%               The following are also provided as short-cuts, with less
+%               overhead and thus better performance on small systems, as
+%               alternatives for 'LU' with the following options:
+%                           OPT.lu.nout     OPT.lu.vec      OPT.lu.thresh
+%                   'LU3'       3               1               1.0
+%                   'LU3a'      3               1
+%                   'LU4'       4               1
+%                   'LU5'       5               1
+%                   'LU3m'      3               0               1.0
+%                   'LU3am'     3               0
+%                   'LU4m'      4               0
+%                   'LU5m'      5               0
 %           'PARDISO' - PARDISO
-%       OPT    : struct of options, with the following fields
-%                (currently used only by PARDISO, default shown in parens,
-%                 see PARDISO documentation for details)
-%           verbose (0) - true or false
-%           mtype (11)  - matrix type (default is real and nonsymmetric)
-%           solver (0)  - solver method (default is sparse direct)
-%           iparm ([])  - n x 2 matrix of integer parameters
-%               1st, 2nd columns are index, value of parameter respectively
-%           dparm ([])  - n x 2 matrix of double parameters
-%               1st, 2nd columns are index, value of parameter respectively
-
-%   Note: SOLVER can also take value 'LU_GP0' or 'LU_AMD0', both of which
-%         just use permutation matrices instead of permutation vectors,
-%         which is slightly slower.
-%         Included for compatibility with Matlab < 7.3.
+%       OPT    : struct of options for certain solvers (e.g. LU and PARDISO)
+%           lu : struct of options to determine form of call to LU solver,
+%                 with the following possible fields (default value in parens):
+%               nout (4) - number of output args for call to LU, UMFPACK is
+%                   used for 4 or 5 output args, Gilbert-Peierls algorithm
+%                   with AMD ordering for 3 output args.
+%               vec (1)  - use permutation vectors instead of matrices
+%                   (permutation matrices used by default for Matlab < 7.3)
+%               thresh   - pivot threshold, see 'help lu' for details
+%           pardiso : struct of PARDISO options (default shown in parens),
+%                 see PARDISO documentation for details
+%               verbose (0) - true or false
+%               mtype (11)  - matrix type (default is real and nonsymmetric)
+%               solver (0)  - solver method (default is sparse direct)
+%               iparm ([])  - n x 2 matrix of integer parameters
+%                   1st, 2nd columns are index, value of parameter respectively
+%               dparm ([])  - n x 2 matrix of double parameters
+%                  1st, 2nd columns are index, value of parameter respectively
 
 %   MIPS
 %   Copyright (c) 2015-2017, Power Systems Engineering Research Center (PSERC)
@@ -54,38 +65,98 @@ end
 
 info = [];
 
-switch upper(solver)
-    case {'', '\'}      %% use built-in \ operator
+switch solver
+    case {'', '\'}  %% use built-in \ operator
         x = A \ b;
-    case {'LU'}         %% explicit LU (is this what's done internally by '\'?)
+    case 'LU3'      %% 3 output LU: Gilbert-Peierls alg, perm vec, 1.0 piv thresh
+        q = amd(A);     %% permutation vector for AMD reordering
+        [L, U, p] = lu(A(q,q), 1.0, 'vector');
+        x = zeros(size(A, 1), 1);
+        x(q) = U \ ( L \ b(q(p)) );
+    case 'LU3a'     %% 3 output LU: Gilbert-Peierls alg, permutation vectors
+        q = amd(A);     %% permutation vector for AMD reordering
+        [L, U, p] = lu(A(q,q), 'vector');
+        x = zeros(size(A, 1), 1);
+        x(q) = U \ ( L \ b(q(p)) );
+    case 'LU4'      %% 4 output LU: UMFPACK, permutation vectors
         [L, U, p, q] = lu(A, 'vector');
         x = zeros(size(A, 1), 1);
         x(q) = U \ ( L \ b(p));
-    case {'LU_GP', 'LU_AMD', 'LU_GP0', 'LU_AMD0'}   %% LU with AMD reordering
-        q = amd(A);                 %% permutation vector for AMD reordering
-        n = size(A, 1);
-        if have_lu_vec() && solver(end) ~= '0'
-            if solver(4) == 'G'     %% Gilbert-Peierls LU
-                %% set pivot threshold to 1 (partial pivoting)
-                %% to ensure that Gilbert-Peierls is used
-                %% (thanks to Jose Luis Marin, see also T. Davis)
-                [L, U, p] = lu(A(q,q), 1.0, 'vector');
-            else
-                [L, U, p] = lu(A(q,q), 'vector');
+    case 'LU5'      %% 5 output LU: UMFPACK w/row scaling, permutation vectors
+        [L, U, p, q, R] = lu(A, 'vector');
+        x = zeros(size(A, 1), 1);
+        x(q) = U \ ( L \ (R(:, p) \ b));
+    case 'LU3m'     %% 3 output LU: Gilbert-Peierls alg, perm mat, 1.0 piv thresh
+        Q = sparse(amd(A), 1:size(A, 1), 1);    %% permutation matrix for AMD reordering
+        [L, U, P] = lu(Q'*A*Q, 1.0);
+        x = Q * ( U \ ( L \ (P * Q' * b)) );
+    case 'LU3am'    %% 3 output LU: Gilbert-Peierls alg, permutation matrices
+        Q = sparse(amd(A), 1:size(A, 1), 1);  %% permutation matrix for AMD reordering
+        [L, U, P] = lu(Q'*A*Q);
+        x = Q * ( U \ ( L \ (P * Q' * b)) );
+    case 'LU4m'     %% 4 output LU: UMFPACK, permutation matrices
+        [L, U, P, Q] = lu(A);
+        x = Q * (U \ (L \ (P * b)));
+    case 'LU5m'     %% 5 output LU: UMFPACK w/row scaling, permutation matrices
+        [L, U, P, Q, R] = lu(A);
+        x = Q * (U \ (L \ (P * (R \ b))));
+    case 'LU'       %% explicit LU, with options struct
+        %% default options
+        nout = 4;               %% 4 output args, UMFPACK
+        vec = have_lu_vec();    %% use permulation vectors, if available
+        thresh = [];            %% use default pivot threshold
+        if isfield(opt, 'lu')
+            opt_lu = opt.lu;
+            if isfield(opt_lu, 'nout')
+                nout = opt_lu.nout;
             end
-            x = zeros(n, 1);
-            x(q) = U \ ( L \ b(q(p)) );
-        else
-            Q = sparse(q, 1:n, 1);  %% permutation matrix for AMD reordering
-            if solver(4) == 'G'     %% Gilbert-Peierls LU
-                %% set pivot threshold to 1 (partial pivoting)
-                %% to ensure that Gilbert-Peierls is used
-                %% (thanks to Jose Luis Marin, see also T. Davis)
-                [L, U, P] = lu(Q'*A*Q, 1.0);
-            else
-                [L, U, P] = lu(Q'*A*Q);
+            if isfield(opt_lu, 'vec')
+                vec = opt_lu.vec;
             end
-            x = Q * ( U \ ( L \ (P * Q' * b)) );
+            if isfield(opt_lu, 'thresh')
+                thresh = opt_lu.thresh;
+            end
+        end
+        %% call the appropriate form
+        switch nout
+            case 3      %% 3 output args: Gilbert-Peierls algorithm, with AMD reordering
+                q = amd(A);         %% permutation vector for AMD reordering
+                n = size(A, 1);
+                if vec
+                    if isempty(thresh)
+                        [L, U, p] = lu(A(q,q), 'vector');
+                    else
+                        [L, U, p] = lu(A(q,q), thresh, 'vector');
+                    end
+                    x = zeros(n, 1);
+                    x(q) = U \ ( L \ b(q(p)) );
+                else
+                    Q = sparse(q, 1:n, 1);  %% permutation matrix for AMD reordering
+                    if isempty(thresh)
+                        [L, U, P] = lu(Q'*A*Q);
+                    else
+                        [L, U, P] = lu(Q'*A*Q, thresh);
+                    end
+                    x = Q * ( U \ ( L \ (P * Q' * b)) );
+                end
+            case 4      %% 4 output args: UMFPACK
+                if vec
+                    [L, U, p, q] = lu(A, 'vector');
+                    x = zeros(size(A, 1), 1);
+                    x(q) = U \ ( L \ b(p));
+                else
+                    [L, U, P, Q] = lu(A);
+                    x = Q * (U \ (L \ (P * b)));
+                end
+            case 5      %% 5 output args: UMFPACK w/row scaling
+                if vec
+                    [L, U, p, q, R] = lu(A, 'vector');
+                    x = zeros(size(A, 1), 1);
+                    x(q) = U \ ( L \ (R(:, p) \ b));
+                else
+                    [L, U, P, Q, R] = lu(A);
+                    x = Q * (U \ (L \ (P * (R \ b))));
+                end
         end
     case {'PARDISO'}
         %% get number of threads from OpenMP env variable
@@ -103,27 +174,27 @@ switch upper(solver)
         solver = 0;
 
         %% override if provided via opt
-        if ~isempty(opt)
-            if isfield(opt, 'verbose')
-                verbose = opt.verbose;
+        if ~isempty(opt) && isfield(opt, 'pardiso')
+            if isfield(opt.pardiso, 'verbose')
+                verbose = opt.pardiso.verbose;
             end
-            if isfield(opt, 'mtype')
-                mtype = opt.mtype;
+            if isfield(opt.pardiso, 'mtype')
+                mtype = opt.pardiso.mtype;
             end
-            if isfield(opt, 'solver')
-                solver = opt.solver;
+            if isfield(opt.pardiso, 'solver')
+                solver = opt.pardiso.solver;
             end
         end
 
         %% begin setup and solve
         info = pardisoinit(mtype, solver);
         info.iparm(3) = num_threads;
-        if ~isempty(opt)
-            if isfield(opt, 'iparm') && ~isempty(opt.iparm)
-                info.iparm(opt.iparm(:, 1)) = opt.iparm(:, 2);
+        if ~isempty(opt) && isfield(opt, 'pardiso')
+            if isfield(opt.pardiso, 'iparm') && ~isempty(opt.pardiso.iparm)
+                info.iparm(opt.pardiso.iparm(:, 1)) = opt.pardiso.iparm(:, 2);
             end
-            if isfield(opt, 'dparm') && ~isempty(opt.dparm)
-                info.iparm(opt.dparm(:, 1)) = opt.dparm(:, 2);
+            if isfield(opt.pardiso, 'dparm') && ~isempty(opt.pardiso.dparm)
+                info.iparm(opt.pardiso.dparm(:, 1)) = opt.pardiso.dparm(:, 2);
             end
         end
         info = pardisoreorder(A, info, verbose);